We improve strategy for getting basemirror for mmdebstrap
Previously we tried to mitigate upgrade issue for essential packages
by selecting more recent timestamp in dependencies. Instead of doing
so and if available, we try to find build-essential or util-linux
dependency timestamp. Else, we fallback to previously using older
timestamp. This is only in the case we are not using metasnap.